Finally adding maze movement infrastructure
Adding grid layout and buttons for moving around the maze, gridview for
the maze, and onclicklistener for moving around. Also added some
try-catch blocks to MainActivity to gain a bit more information when the
application crashes in the emulator. Next steps are for me to implement
moving around the maze using the onclicklistener.