File Path Conclusions and Activity Results
*Added away to return data from the roadblock activity
*Added a close button to the roadblock ui
*Added a couple ways to interact with files [1] (via /res [2])  (/assets
[3]) in the mazegui -> roadblock encounter function

[1]
After some exhaustive testing, the final word on file paths are that
they're not going to happen easily. All the files we place inside the
application package aren't accessible via the File class. Therefore the
only way to access these files (write/read) are to use two different
options. One is the /res folder, the other is the /assets folder. Each
one has it's pros and cons.

[2] The /res folder basically interacts with the android OS. Since we're
using an xml file, it's compatible with the OS features and gets
compiled. The OS generates a resource identifier and which makes working
with the file pretty easy. The only problem is that there may be some
unforeseen bugs.

[3] /assets folder is for any kind of data file we want to add. This
includes images, movies, sounds, and any every other file we'll need.
Interacting with the files inside the assets folder is done via the
Assets manager. Since no id is made, you'll have to write the file path
and extension when opening the stream.