Movement Implementation
Committing movement implementation. Added some logging/exception
handling. Movement should update user_coords in the maze generation
object and switch to a roadblock when 'R' is encountered. Currently I'm
getting invalid move every time and I think this relates to generating
the maze and accessing the maze through the maze generation object.
Next to implement is handling when the maze is solved.