Added Player Icon, Post-movement Maze Updates
Note: The direction scheme of my maze graphics seems to not match that
of Erika's movement class, so I slightly altered her button bindings in
the maze GUI file to compensate.

Additionally, I added an array that holds the element ids of the
individual cells in the mazeImage GridLayout. When trying to access a
cell ImageView by id, use the cell's coordinates in the idArray[][] to
get the id.